fixed weapon update, made ammo ints

+++ CubeFace/src/com/zaxxon/world/shooting/Weapon.java
@@ -52,8 +52,8 @@ public class Weapon {
 	 * @return ammo
 	 */
 	public String getAmmo() {
-		
-		return Double.toString(currentAmmo);
+
+		return Integer.toString(((int) currentAmmo));
 	}
 	
 	/**
